minor refactor for consistency
