+++ client/src/actions/traffic.js
@@ -96,34 +96,3 @@ export const fetchTraffic = (dateString) => {
     dispatchToTimeline(dispatch, getState, dateString, response);
   }
 };
-
-export const fetchPaths = () => {
-  return async (dispatch) => {
-    const response = await resource.fetchPaths();
-
-    dispatch({
-      type: FETCH_PATHS,
-      payload: response.data
-    });
-  }
-};
-
-export const fetchPathDetail = (from, to, timestamp) => {
-  return async (dispatch) => {
-    dispatch({
-      type: FETCH_PATH_DETAIL_INITIALIZED,
-      payload: {
-        from,
-        to,
-        timestamp
-      }
-    });
-
-    const response = await resource.fetchPathDetail(from, to, timestamp);
-    dispatch({
-      type: FETCH_PATH_DETAIL,
-      payload: response.data
-    });
-  };
-};
-
