+++ client/src/reducers/index.js
@@ -2,11 +2,13 @@ import { combineReducers } from 'redux';
 import traffic from './trafficReducer';
 import timeline from './timelineReducer';
 import route from './routeReducer';
+import path from './pathReducer';
 
 const rootReducer = combineReducers({
   traffic,
   timeline,
-  route
+  route,
+  path
 });
 
 export default rootReducer;
