+++ client/src/reducers/trafficReducer.js
@@ -1,14 +1,9 @@
 import {
   FETCH_TRAFFIC,
-  FETCH_PATHS,
-  FETCH_PATH_DETAIL_INITIALIZED,
-  FETCH_PATH_DETAIL
 } from '../actions/types';
 
 const initialState = {
-  trafficByTimestamp: {},
-  paths: [],
-  selectedPathDetail: null
+  trafficByTimestamp: {}
 };
 
 const trafficReducer = (state = initialState, action) => {
@@ -25,21 +20,6 @@ const trafficReducer = (state = initialState, action) => {
       return {
         ...state,
         trafficByTimestamp
-      };
-    case FETCH_PATHS:
-      return {
-        ...state,
-        paths: action.payload.paths || []
-      };
-    case FETCH_PATH_DETAIL_INITIALIZED:
-      return {
-        ...state,
-        selectedPathDetail: null
-      };
-    case FETCH_PATH_DETAIL_INITIALIZED:
-      return {
-        ...state,
-        selectedPathDetail: action.payload.pathDetail
       };
     default:
       return state;
